Fix: properly support 26.1.x in FMJ, move non-mixin classes out of mixin package, revert client language collection to non-codecs for better performance
Also rename MinecraftAssetResolver to ClientAssetResolver
