+++ datagen/src/main/java/org/geysermc/rainbow/datagen/RainbowDataProvider.java
@@ -33,8 +33,8 @@ import org.geysermc.rainbow.Rainbow;
 import org.geysermc.rainbow.RainbowIO;
 import org.geysermc.rainbow.datagen.mixin.DataComponentInitializersAccessor;
 import org.geysermc.rainbow.datagen.mixin.FabricLanguageProviderAccessor;
-import org.geysermc.rainbow.datagen.mixin.LanguageProviderDataAccessor;
-import org.geysermc.rainbow.datagen.mixin.ModelProviderDataAccessor;
+import org.geysermc.rainbow.datagen.accessor.LanguageProviderDataAccessor;
+import org.geysermc.rainbow.datagen.accessor.ModelProviderDataAccessor;
 import org.geysermc.rainbow.mapping.AssetResolver;
 import org.geysermc.rainbow.mapping.PackSerializer;
 import org.geysermc.rainbow.mapping.texture.TextureResource;
