+++ datagen/src/main/java/org/geysermc/rainbow/datagen/accessor/LanguageProviderDataAccessor.java
@@ -1,4 +1,4 @@
-package org.geysermc.rainbow.datagen.mixin;
+package org.geysermc.rainbow.datagen.accessor;
 
 import java.util.Map;
