+++ datagen/src/main/java/org/geysermc/rainbow/datagen/mixin/ModelProviderMixin.java
@@ -8,6 +8,7 @@ import net.minecraft.data.CachedOutput;
 import net.minecraft.data.DataProvider;
 import net.minecraft.resources.Identifier;
 import net.minecraft.world.item.Item;
+import org.geysermc.rainbow.datagen.accessor.ModelProviderDataAccessor;
 import org.jspecify.annotations.Nullable;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Unique;
